Closes-Bug: #1470624 Server Manager: Pupept Apache module should not remove existing configs
Fix Done:
Add a virtual host for apache to listen on 9003,
so that we dont overlap with existing default virtual host of apache.

Also when we install openstack role on same machine as smgr,
the horizon vhost conf takes precedence over the default.

So we explicity create a separate virtual host for smgr
related stuff.